Revert "Math fixes, and inclusion of steering angle into visualization"
This reverts commit f6c5a7c82beda40b38cb0d94e0c42ca22329b1ad.

+++ detector/interface/proj.h
@@ -114,7 +114,7 @@ struct OutputContainer
 	cv::Point2f BezPointOne;
 	cv::Point2f BezPointTwo;
 	cv::Point2f BezPointThree;
-	double SteerAngle;
+	float SteerAngle;
 };
